refactor: extract CloseDialog logic into useCloseLease composable
CloseDialog.vue had ~776 script LOC (3 watchers, a debounce timer, an
inject, dialog orchestration) inside one SFC, against the house rule that
a template be a thin projection of already-computed values (issue #185,
this is the first file in the epic).

Move all data/derivation/validation/swap-fee/submit logic verbatim into a
co-located useCloseLease() composable, and the preview block into a
props-only ClosePreview.vue sub-component. CloseDialog.vue drops from
~1066 to ~53 lines and now only wires the composable to the template.

Behaviour is unchanged: the existing CloseDialog.test.ts (13 cases,
hoisted mocks + debounce timers) passes untouched and exercises the moved
paths transitively. The 600ms setSwapFee debounce floor and its rate-limit
rationale, the onUnmounted timer teardown, and every watcher's
dependency/idempotency contract move with their code.

One line moved from a template expression into .ts script changed `==` to
`===` (route.matched path check) to satisfy the eqeqeq lint rule, which
does not apply inside .vue template expressions; both operands are strings
so the comparison is identical.
